Homework1/requirements.txt
torch==1.5.1
torchvision==0.6.1
numpy
pillow>=4.1.1
